requirements.txt
django~=5.0.2
djangorestframework~=3.14.0
djangorestframework-jwt~=1.11.0
djangorestframework-simplejwt